cleanup & tweaks to avoid UnsatisfiableErrors
